+++ src/bin/keyring-cli.rs
@@ -18,13 +18,13 @@
 #[macro_use]
 extern crate log;
 
-use amplify::Exec;
 use clap::Clap;
 use log::LevelFilter;
 use std::convert::TryInto;
 
 use keyring::cli::{Config, Opts, Runtime};
 use keyring::error::BootstrapError;
+use keyring::Exec;
 
 #[tokio::main]
 async fn main() -> Result<(), BootstrapError> {