+++ CHANGELOG.md
@@ -1,1 +1,11 @@
-# Change Log
+Change Log
+==========
+
+v0.1.0-beta.1
+-------------
+
+Initial release
+- Keyring and key accounts creation
+- Key derivation
+- Basic RPC API
+- Message signatures